Just optimized my architecture of the project to be clean architecture
