+++ lib/modules/auth/presentation/views/forget_password_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../../../core/utils/app_strings.dart';
-import '../../../core/utils/app_text_styles.dart';
-import '../../../core/widgets/custom_header_text.dart';
+import '../../../../core/utils/app_strings.dart';
+import '../../../../core/utils/app_text_styles.dart';
+import '../../../../core/widgets/custom_header_text.dart';
 import '../widgets/custom_forget_password_form_widget.dart';
 
 class ForgetPasswordView extends StatelessWidget {
