+++ lib/modules/auth/presentation/views/login_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import '../../../core/functions/navigation.dart';
-import '../../../core/utils/app_strings.dart';
-import '../../../core/utils/app_text_styles.dart';
+import '../../../../core/functions/navigation.dart';
+import '../../../../core/utils/app_strings.dart';
+import '../../../../core/utils/app_text_styles.dart';
 import '../widgets/custom_have_account_widget.dart';
 import '../widgets/custom_login_form_widget.dart';
 import '../widgets/custom_or_divider_widget.dart';
