+++ lib/modules/auth/presentation/views/reset_your_password_view.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 
-import '../../../core/utils/app_strings.dart';
-import '../../../core/utils/app_text_styles.dart';
-import '../../../core/widgets/custom_header_text.dart';
+import '../../../../core/utils/app_strings.dart';
+import '../../../../core/utils/app_text_styles.dart';
+import '../../../../core/widgets/custom_header_text.dart';
 
-class PasswordRecoveryView extends StatelessWidget {
-  const PasswordRecoveryView({super.key});
-  static const String routeName = "/passwordRecovery-View";
+class ResetyourpasswordView extends StatelessWidget {
+  const ResetyourpasswordView({super.key});
+  static const String routeName = "/resetPassword-view";
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -18,7 +18,7 @@ class PasswordRecoveryView extends StatelessWidget {
             const SliverToBoxAdapter(child: SizedBox(height: 60)),
             SliverToBoxAdapter(
               child: CustomHeaderText(
-                text1: AppStrings.checkYourEmailTitle,
+                text1: AppStrings.newPassword,
                 textAlign1: TextAlign.center,
                 styleText1: AppTextStyle.Cairo700style23.copyWith(fontSize: 19),
               ),
@@ -26,7 +26,7 @@ class PasswordRecoveryView extends StatelessWidget {
             SliverToBoxAdapter(child: SizedBox(height: 16)),
             SliverToBoxAdapter(
               child: CustomHeaderText(
-                text1: AppStrings.checkYourNumberSubTitle,
+                text1: AppStrings.createNewPassword,
                 textAlign1: TextAlign.start,
                 styleText1: AppTextStyle.Cairo600style16,
               ),
@@ -38,4 +38,3 @@ class PasswordRecoveryView extends StatelessWidget {
     );
   }
 }
-
