+++ lib/modules/auth/presentation/widgets/custom_checkbox_widget.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 
-import '../../../../core/utils/app_colors.dart';
+import '../../../../../core/utils/app_colors.dart';
 
 class CustomCheckBox extends StatefulWidget {
   const CustomCheckBox({super.key});
