+++ lib/modules/auth/presentation/widgets/custom_forget_password_form_widget.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:fruits_hup/core/functions/navigation.dart';
-import 'package:fruits_hup/modules/auth/views/password_recovery_view.dart';
+import 'package:fruits_hup/modules/auth/presentation/views/password_recovery_view.dart';
 
-import '../../../core/functions/validation.dart';
-import '../../../core/utils/app_strings.dart';
-import '../../../core/utils/app_text_styles.dart';
-import '../../../core/widgets/custom_btn.dart';
+import '../../../../core/functions/validation.dart';
+import '../../../../core/utils/app_strings.dart';
+import '../../../../core/utils/app_text_styles.dart';
+import '../../../../core/widgets/custom_btn.dart';
 import 'cutom_text_form_field_widget_.dart';
 
 class CustomForgetPasswordForm extends StatelessWidget {
