+++ lib/modules/auth/presentation/widgets/custom_have_account_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:fruits_hup/core/utils/app_colors.dart';
-import '../../../../core/utils/app_text_styles.dart';
+import '../../../../../core/utils/app_text_styles.dart';
 
 class HaveAccountWidget extends StatelessWidget {
   const HaveAccountWidget({
