+++ lib/modules/auth/presentation/widgets/custom_or_divider_widget.dart
@@ -2,8 +2,8 @@
 
 import 'package:flutter/material.dart';
 
-import '../../../core/utils/app_colors.dart';
-import '../../../core/utils/app_text_styles.dart';
+import '../../../../core/utils/app_colors.dart';
+import '../../../../core/utils/app_text_styles.dart';
 
 class CustomOrDivider extends StatelessWidget {
   CustomOrDivider({super.key, required this.text});
