+++ lib/modules/auth/presentation/widgets/custom_terms_and_conds_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:fruits_hup/core/utils/app_colors.dart';
 
-import '../../../../core/utils/app_strings.dart';
-import '../../../../core/utils/app_text_styles.dart';
+import '../../../../../core/utils/app_strings.dart';
+import '../../../../../core/utils/app_text_styles.dart';
 import 'custom_checkbox_widget.dart';
 
 class TermsAndConditionsWidget extends StatelessWidget {
